Camera game, now can be setted, and don't create lag when eliminated

+++ Kebab Engine/Source/ComponentCamera.h
@@ -61,8 +61,8 @@ private:
 	float fovVertical;
 	float fovHorizontal;
 
-	float currentWinWidth;
-	float currentWinHeight;
+	int currentWinWidth;
+	int currentWinHeight;
 
 	float planeFar;
 	float planeNear;
